Rename listHandler to nextHandler
